project/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
pip==22.0.4
pytz==2023.3
setuptools==58.1.0
sqlparse==0.4.3
tzdata==2023.3